ingester_app/requirements.txt
Flask>=2.0 
psycopg2-binary>=2.9 
python-dotenv>=0.19 